requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.20.0
Flask-Migrate==3.1.0
Werkzeug==2.0.3

